Rename `language` to `languages` directory
