+++ packages/langium/src/grammar/type-system/type-collector/all-types.ts
@@ -4,7 +4,7 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import type { ParserRule, Interface, Type, Grammar } from '../../../language/generated/ast.js';
+import type { ParserRule, Interface, Type, Grammar } from '../../../languages/generated/ast.js';
 import type { URI } from '../../../utils/uri-util.js';
 import type { LangiumDocuments } from '../../../workspace/documents.js';
 import type { PlainAstTypes } from './plain-types.js';
@@ -12,7 +12,7 @@ import type { AstTypes } from './types.js';
 import { collectInferredTypes } from './inferred-types.js';
 import { collectDeclaredTypes } from './declared-types.js';
 import { getDocument } from '../../../utils/ast-util.js';
-import { isParserRule } from '../../../language/generated/ast.js';
+import { isParserRule } from '../../../languages/generated/ast.js';
 import { resolveImport } from '../../internal-grammar-util.js';
 import { isDataTypeRule } from '../../../utils/grammar-util.js';
