+++ packages/langium/src/languages/grammar-config.ts
