+++ packages/langium/src/languages/language-meta-data.ts
