+++ packages/langium/src/lsp/hover-provider.ts
@@ -5,7 +5,7 @@
  ******************************************************************************/
 
 import type { CancellationToken, Hover, HoverParams } from 'vscode-languageserver';
-import type { GrammarConfig } from '../language/grammar-config.js';
+import type { GrammarConfig } from '../languages/grammar-config.js';
 import type { References } from '../references/references.js';
 import type { LangiumServices } from '../services.js';
 import type { AstNode } from '../syntax-tree.js';
