+++ packages/langium/src/syntax-tree.ts
@@ -6,7 +6,7 @@
 
 import type { TokenType } from 'chevrotain';
 import type { URI } from './utils/uri-util.js';
-import type { AbstractElement } from './language/generated/ast.js';
+import type { AbstractElement } from './languages/generated/ast.js';
 import type { DocumentSegment, LangiumDocument } from './workspace/documents.js';
 
 /**
